requirements.txt
lxml>=4.9.1
openpyxl>=3.1.2
pandas>=1.3.5
pytest>=7.1.2
python_docx>=0.8.11
python_pptx>=0.6.21
tabulate
filetype
